Create generate_org_chart.sql
This query generates an org chart, in text format (as a series of rows listing employee name, department, and the name of each employee's manager. The reason for the common table expression (cte) query is to generate the level of each employee in the org chart, where the person at the top of the organization is at Level 0.